ARM: shmobile: kzm9g-reference dts: Add chosen/stdout-path
Add a stdout-path property so that automatic console selection works
in the absence of a "console=" parameter on the kernel command line.
In addition, this will allow the PM domain code to find the PM domain
for the console device.

Note that we have to keep the "console=ttySC4" parameter in
chosen/bootargs, else we only get console messages on tty0.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
(cherry picked from commit 59ab5bff23020b6223857a460a39b9ec71d91cea)
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>